created separate file for genreList
