+++ src/Routes/Movies/index.js
@@ -5,7 +5,8 @@ import {
 } from "./styled";
 import star from "../../Images/star.svg";
 import { useSelector } from "react-redux";
-import { selectGenreList, selectGenreListState, selectMovieList, selectMovieListState } from "../../features/moveList/movieListSlice";
+import { selectMovieList, selectMovieListState } from "../../features/moveList/movieListSlice";
+import { selectGenreList, selectGenreListState } from "../../features/genreList/genreListSlice";
 
 const Movies = () => {
     const fetchData = useSelector(selectMovieList);
@@ -13,12 +14,16 @@ const Movies = () => {
     const loadingGeners = useSelector(selectGenreListState);
     const loadingMovies = (useSelector(selectMovieListState));
     const URL = "https://www.themoviedb.org/t/p/w440_and_h660_face/";
+    console.log(fetchData);
+        console.log(fetchGenre);
 
-    if (loadingMovies.status === "success" && loadingGeners.statusGenre === "success") {
+    if (loadingMovies.status === "success" && loadingGeners.status === "success") {
         const moviesList = fetchData.results;
         const genreList = fetchGenre.genres;
         const genres = [];
         
+        console.log(moviesList);
+        console.log(genreList);
         return (
             <>
                 <StyledTitle>
