+++ src/features/moveList/MovieList/index.js
@@ -1,5 +1,6 @@
 import { useDispatch } from "react-redux";
-import { axiosMovieList, axiosGenreList } from "../movieListSlice";
+import { axiosMovieList} from "../movieListSlice";
+import { axiosGenreList } from "../../genreList/genreListSlice";
 import { useEffect } from "react";
 
 export const MovieList = () => {
